Remove verbal codes
