+++ package.json
@@ -12,7 +12,7 @@
         "build:noui": "lerna run --stream --ignore itmat-docker --ignore itmat-ui-react build",
         "dockerise": "yarn run build:noui && lerna run --stream --scope itmat-docker build",
         "setupDatabase": "lerna run --scope itmat-setup setup",
-        "test": "lerna run --no-bail --concurrency 1 --stream --scope itmat-interface --scope itmat-job-executor --scope itmat-commons test",
+        "test": "SKIP_EMAIL_TEST=truelerna run --no-bail --concurrency 1 --stream --scope itmat-interface --scope itmat-job-executor --scope itmat-commons test",
         "lint": "lerna run --no-bail --stream lint",
         "lint:fix": "lerna run --no-bail --stream lint:fix",
         "clean": "lerna run clean",
