+++ packages/itmat-interface/src/graphql/schemaV0.ts
@@ -1,6 +1,6 @@
 import { gql } from 'apollo-server-express';
 
-export const schemaV2 = gql`
+export const schemaV0 = gql`
 scalar JSON
 
 enum USERTYPE {
@@ -167,78 +167,6 @@ type Job {
     data: JSON
 }
 
-enum LOG_TYPE {
-   SYSTEM_LOG
-   REQUEST_LOG
-}
-
-enum LOG_STATUS {
-    SUCCESS
-    FAIL
-}
-
-enum LOG_ACTION {
-    # SYSTEM
-    START_SERVER
-    STOP_SERVER
-
-    # USER
-    GET_USERS
-    EDIT_USER
-    DELETE_USER
-    CREATE_USER
-    LOGIN_USER
-    WHO_AM_I
-    LOGOUT
-    REQUEST_USERNAME_OR_RESET_PASSWORD
-    RESET_PASSWORD
-
-    # PROJECT
-    GET_PROJECT
-    # GET_PROJECT_PATIENT_MAPPING
-    EDIT_PROJECT_APPROVED_FIELDS
-    EDIT_PROJECT_APPROVED_FILES
-    CREATE_PROJECT
-    DELETE_PROJECT
-    SET_DATAVERSION_AS_CURRENT
-    SUBSCRIBE_TO_JOB_STATUS
-
-    # STUDY | DATASET
-    DELETE_STUDY
-    GET_STUDY
-    GET_STUDY_FIELDS
-    CREATE_STUDY
-    CREATE_DATA_CREATION_JOB
-    #CREATE_FIELD_CURATION_JOB
-
-    # STUDY & PROJECT
-    EDIT_ROLE
-    ADD_NEW_ROLE
-    REMOVE_ROLE
-
-    # FILE
-    UPLOAD_FILE
-    DOWNLOAD_FILE
-    DELETE_FILE
-
-    #QUERY
-    GET_QUERY
-    CREATE_QUERY
-    #GET_QUERY_RESULT
-}
-
-type Log {
-    id: String!,
-    requesterName: String,
-    requesterType: USERTYPE,
-    logType: LOG_TYPE,
-    actionType: LOG_ACTION,
-    actionData: JSON,
-    time: Float!,
-    status: LOG_STATUS,
-    error: String
-}
-
 type QueryEntry {
     id: String!,
     queryString: String!,
@@ -339,8 +267,6 @@ type Query {
     # PERMISSION
     getGrantedPermissions(studyId: String, projectId: String): UserPermissions
 
-    # LOG
-    getLogs(requesterName: String, requesterType: USERTYPE, logType: LOG_TYPE, actionType: LOG_ACTION, status: LOG_STATUS): [Log]
 }
 
 type Mutation {
