+++ packages/itmat-ui-react/src/components/apolloClient.tsx
@@ -4,7 +4,7 @@ import { onError } from '@apollo/client/link/error';
 import { createUploadLink } from 'apollo-upload-client';
 
 const uploadLink = createUploadLink({
-    uri: `${window.location.origin}/api/v2`,
+    uri: `${window.location.origin}/api/v1`,
     credentials: 'include'
 }) as any as ApolloLink;
